+++ CHANGELOG.md
@@ -13,6 +13,7 @@ changes.
 ### Added
 
 - Preserve maintenance ending banner state on the wallet connection change [Issue 3681](https://github.com/IntersectMBO/govtool/issues/3681)
+- Add authors for Live Voting Governance Actions [Issue 3745](https://github.com/IntersectMBO/govtool/issues/3745)
 
 ### Fixed